WS-034: Refactor buttons into individual components
