+++ frontend/src/components/SightingListPage/AdminButtons/ConfirmButton.tsx
@@ -1,32 +1,21 @@
 import React, { useContext } from "react";
-import {
-  approveSighting,
-  deleteSighting,
-  User,
-} from "../../../clients/apiClients";
+import { approveSighting, User } from "../../../clients/apiClients";
 import { LoginContext } from "../../login/LoginManager";
 
-export function AdminButtons({
+export function ConfirmButton({
   approvedBy,
   sightingId,
   setActionOnConfirm,
-  setActionOnDelete,
 }: {
   approvedBy: User;
   sightingId: number;
   setActionOnConfirm: React.Dispatch<React.SetStateAction<boolean>>;
-  setActionOnDelete: React.Dispatch<React.SetStateAction<boolean>>;
 }): JSX.Element {
   const { username, password } = useContext(LoginContext);
 
   const confirmWhaleSighting = () => {
     approveSighting(sightingId, username, password).then(() =>
       setActionOnConfirm(true)
-    );
-  };
-  const deleteWhaleSighting = () => {
-    deleteSighting(sightingId, username, password).then(() =>
-      setActionOnDelete(true)
     );
   };
 
@@ -41,15 +30,6 @@ export function AdminButtons({
         type="submit"
       >
         Confirm
-      </button>
-      <button
-        className="sighting__button btn btn-primary"
-        onClick={() => {
-          deleteWhaleSighting();
-        }}
-        type="submit"
-      >
-        Delete
       </button>
     </div>
   );
